+++ releases.md
@@ -1,5 +1,11 @@
 # Releases
 
+## Unreleased
+
+  - Add agent context.
+  - Add support for pattern pub/sub.
+  - Add support for sharded pub/sub.
+
 ## v0.11.2
 
   - Fix handling of IPv6 address literals, including those returned by Redis Cluster / Sentinel.